Fix #56, undo uniqueid patch
The `uniqueid` module had a bug when "imported" with the RepyV1
preprocessor, `repypp`. (That program had no notion of modules already
imported, so it would paste multiple copies of imported libraries
and thus cause duplicate IDs with the `uniqueid` module). This bug
was patched for V1, but the patch never removed since.

In RepyV2, `dylink` is used to import modules, and it does a much
better job at importing modules only once, see for instance
SeattleTestbed/seattlelib_v2#55. Therefore, we can remove the old
patch!

To verify that removing the patch is safe, I also added a unit test
that checks `uniqueid`'s
* basic functionality,
* functioning after reimport, and
* function across threads.